refactor: move integracao module to core
