+++ central-oon-core-backend/src/errors/integracao/integracaoNaoEncontrado.js
@@ -1,4 +1,4 @@
-const GenericError = require("../generic");
+const GenericError = require("../GenericError");
 
 class IntegracaoNaoEncontradoError extends GenericError {
   constructor() {
