+++ central-oon-core-backend/src/models/IntegracaoConfig.js
