+++ src/services/arquivo/omie/central-omie/handler.js
@@ -5,7 +5,7 @@ const { processarIntegracao } = require("../../../queue/defaultHandler.js");
 const {
   fileHandler: { compactFile },
 } = require("central-oon-core-backend");
-const Integracao = require("../../../../models/Integracao.js");
+const { Integracao } = require("central-oon-core-backend");
 
 const handler = async (integracao) => {
   return processarIntegracao({
